[refactor] fletを使用して実装した設定画面を別のファイルに分割した
